+++ src/app/dataset-view/additional-components/flows-component/services/dataset-flows.service.spec.ts
@@ -23,11 +23,15 @@ import {
 import { MaybeUndefined } from "src/app/common/app.types";
 import { Account, DatasetFlowType } from "src/app/api/kamu.graphql.interface";
 import { FlowsTableData } from "src/app/common/components/flows-table/flows-table.types";
+import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { LoggedUserService } from "src/app/auth/logged-user.service";
+import { mockAccountDetails } from "src/app/api/mock/auth.mock";
 
 describe("DatasetFlowsService", () => {
     let service: DatasetFlowsService;
     let datasetFlowApi: DatasetFlowApi;
     let toastService: ToastrService;
+    let loggedUserService: LoggedUserService;
     const MOCK_DATASET_ID = "did:odf:fed0100d72fc7a0d7ced1ff2d47e3bfeb844390f18a7fa7e24ced6563aa7357dfa2e8";
     const MOCK_PAGE = 1;
     const MOCK_PER_PAGE = 15;
@@ -38,11 +42,14 @@ describe("DatasetFlowsService", () => {
     beforeEach(() => {
         TestBed.configureTestingModule({
             providers: [Apollo],
-            imports: [ApolloTestingModule, ToastrModule.forRoot()],
+            imports: [ApolloTestingModule, ToastrModule.forRoot(), HttpClientTestingModule],
         });
         service = TestBed.inject(DatasetFlowsService);
         datasetFlowApi = TestBed.inject(DatasetFlowApi);
         toastService = TestBed.inject(ToastrService);
+        loggedUserService = TestBed.inject(LoggedUserService);
+
+        spyOnProperty(loggedUserService, "currentlyLoggedInUser", "get").and.returnValue(mockAccountDetails);
     });
 
     it("should be created", () => {